pdf-extraction/requirements.txt
google-cloud-vision==1.0.0
google-cloud-storage==1.28.1
oauth2client==4.1.3
gspread==3.6.0

# dev
pytest==5.4.3